ibkr-market-data-infra/requirements.txt
ib_insync
influxdb-client
python-dotenv
flask

